scaffolding cambio de rutas con problemas de conección
